Disallow source HLS playlist
The source stream is content directly from the streamer. It may not be
optimal and Twitch partners will have Twitch transcoded playlists in
additition to source. Non-partnered channels return source even if it
has been disabled, as it is the only quality. Some source streams will
studder if the bitrate it high (e.g. 1080p60 4500> kbps).